refactor: gitignore file edit
