+++ backend/pickit/.gitignore
@@ -35,5 +35,3 @@ out/
 
 ### VS Code ###
 .vscode/
-
-application-secret.yml
